requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
httpx==0.27.0
pyyaml==6.0.1
openai==1.30.1
unidiff==0.7.5
python-dotenv==1.0.1
